[Feat] ViewModel Bindings
